利用多线程优化文件识别

+++ akkad-war3/war3-moon/src/main/java/xyz/wongs/drunkard/war3/moon/entity/FileInfo.java
@@ -16,7 +16,7 @@ public class FileInfo extends AbstractEntity<Long> {
 
     private String filePath;
 
-    private Float fileSize;
+    private Long fileSize;
 
     private String suffixName;
